fix: Set accouting combination values.
